UPSTREAM: mfd: cros_ec_dev: Don't advertise junk features on failure
If the feature query fails during this driver's probe, we memset the
features bits to 0, but then we continue to overwrite that with whatever
junk we read back when the feature probing failed. In the case of a SPI
device, it may be complete junk, causing the caller of this function to
think certain features are present when they aren't. Don't copy over the
bad message on failure so we can fail more gracefully.

Fixes: e4244ebddae2 ("platform/chrome: Introduce a new function to check EC features.")
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Acked-by: Benson Leung <bleung@chromium.org>
Signed-off-by: Lee Jones <lee.jones@linaro.org>

BUG=None
TEST=Boot cheza with bad spi, see message failure and don't see random
EC subdevices probe

(cherry picked from commit df7c3bf24f546ab0e5a23b282d34850f2b8eb0d2)
Change-Id: Ibb5be5de46f055ec91d7caea896b5945ebeea941
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1095648
Reviewed-by: Benson Leung <bleung@google.com>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>