FIXUP: UPSTREAM: ASoC: soc-pcm: check symmetry after hw_params
The upstream pick at CL:881656 fixed the patch _not_ to use spaces for
indentation.  This was the incorrect thing to do and causes future
merge conflicts.  Upstream picks should always be exact and shouldn't
incorporate extra fixes.  If upstream is broken, send a patch
upstream.

This patch changes the two relevant lines to be indented with spaces
again which matches how upstream was.

BUG=b:71724897
TEST=Future picks are clean now.

Change-Id: I3bc80fabd4b1eda4665155734b60cea4300839d2
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1095639
Reviewed-by: Ben Zhang <benzh@chromium.org>