CHROMIUM: ASoC: DA7219: For Grunt save and restore state of register on S3
On Grunt platform at S3 the power line to da7219 is cut-off. As a
result the codec is in reset state.
To provide a software workaround, the register state is saved at the
time of suspend and is restored on resume.

BUG=b:109765944
TEST=
play audio
powerd_dbus_suspend
resume
audio continues to play

Change-Id: I3b7f287d67ea5f172ba3810ffd7b7dd2c892044e
Signed-off-by: Akshu Agrawal <akshu.agrawal@amd.com>
Reviewed-on: https://chromium-review.googlesource.com/1109606
Reviewed-by: Simon Glass <sjg@chromium.org>