CHROMIUM: config: build v4l2 into the kernel in all configs
We're building v4l2 as modules on i386 & chromiumos-arm, but building
it into the kernel for all other configs.  Standardize this and build
it into the kernel for everyone.

In practice this shouldn't make too big a difference: every Chromebook
has had a camera, so the v4l2 module was loaded.  Further, i386 isn't
used anywhere anymore, and we're moving to aarch64 kernels, so the
general arm 32-bit config isn't used that much.

BUG=None
TEST=precq passes

Change-Id: Iddd56f1c734259b4ba1276501cb3ba2f489d4de8
Signed-off-by: Mike Frysinger <vapier@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1111601
Reviewed-by: Guenter Roeck <groeck@chromium.org>