UPSTREAM: ALSA: ac97: add codecs devicetree binding
Add a devicetree binding for codecs. This is especially useful if the
AC97 bitclk clock is provided by the codec, as it has to be described in
the devicetree description for the ac97 bus code to aquire it.

Signed-off-by: Robert Jarzmik <robert.jarzmik@free.fr>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
(cherry picked from commit 2225a3e6af78e605119e0f178ae957d860882b1d)

BUG=chromium:1000361
TEST=tested on v4.14 platforms

Change-Id: I2b69cbb6320c4a3e09683e5843adf77c8f07db6e
Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1804861
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>