FROMLIST: BACKPORT: x86/tsc: calibrate tsc only once
During boot tsc is calibrated twice: once in tsc_early_delay_calibrate(),
and the second time in tsc_init().

Rename tsc_early_delay_calibrate() to tsc_early_init(), and rework it so
the calibration is done only early, and make tsc_init() to use the values
already determined in tsc_early_init().

Sometimes it is not possible to determine tsc early, as the subsystem that
is required is not yet initialized, in such case try again later in
tsc_init().

This patch is taken from the series accepted here:
https://lkml.org/lkml/2018/7/19/878

BUG=b:112103715
TEST=Verified that no hang is observed on yorp for 5000+ warm reboots.

Suggested-by: Thomas Gleixner <tglx@linutronix.de>

Change-Id: I5128fa1a57a57174a33cea976a8d03fcf7e734b3
Signed-off-by: Pavel Tatashin <pasha.tatashin@oracle.com>
Signed-off-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1159343
Commit-Ready: Wenkai Du <wenkai.du@intel.com>
Tested-by: Wenkai Du <wenkai.du@intel.com>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-by: Justin TerAvest <teravest@chromium.org>