CHROMIUM: arm64: dts: qcom: add cheza RMTFS memory carveout
This patch adds the memory carveout for the remote filesystem (RMTFS)
buffer for exchanging filesystem data between the appplication processor
and the modem on SDM845.

The exact amount of memory may differ for various boards, based on the
expected supported features, so add this definition on a per-board
basis. The exact range used is arbitrary, but I believe a 2MB minimum is
about right for now.

BUG=b:79668211
TEST=Test rmtfs server on cheza.

Change-Id: I27c238d9b56785ebed6da63ac89939daca0ad840
Signed-off-by: Ben Chan <benchan@chromium.org>
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1119572
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>