CHROMIUM: iwl7000: mvm: don't send keys when entering D3
In the past, we needed to program the keys when entering
D3. This was since we replaced the image. However, now
that there is a single image, this is no longer needed.
Note that RSC is sent separately in a new command.
This solves issues with newer devices that support PN
offload. Since driver re-sent the keys, the PN got zeroed
and the receiver dropped the next packets, until PN
caught up again.

Change-Id: I29f5a762c73efa5afbd4ae7125a2d4016f16d4d6
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 1cf634de5224a06494861a837fce6fd8294d9320