UPSTREAM: drm: add missing ctx argument to plane transitional helpers
In commits:
34a2ab5e0689 ("drm: Add acquire ctx parameter to ->update_plane")
1931529448bc ("drm: Add acquire ctx parameter to ->plane_disable")

a pointer to a drm_modeset_acquire_ctx structure was added as an
argument to the method prototypes.  The transitional helpers are
supposed to be directly plugged in as implementations of these
methods, but doing so generates a warning.  Add the missing
argument.

A number of buggy users were added for drm_plane_helper_disable()
which need to be fixed up for this change, which we do by passing
a NULL ctx argument.

Fixes: 1931529448bc ("drm: Add acquire ctx parameter to ->plane_disable")
Signed-off-by: Russell King <rmk+kernel@armlinux.org.uk>
Signed-off-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Link: https://patchwork.freedesktop.org/patch/msgid/E1fa1Zr-0005gT-VF@rmk-PC.armlinux.org.uk
(cherry picked from commit 070473bcf703366e9acb14e172d5b6563cc07a26)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:80589817
TEST=Tested on cheza

Change-Id: I645ce8fb3ee939594e93a6e8f677df42761d4e4d
Reviewed-on: https://chromium-review.googlesource.com/1158888
Commit-Ready: Sean Paul <seanpaul@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Kristian H. Kristensen <hoegsberg@chromium.org>