FROMGIT: ASoC: AMD: Reset bytescount when starting transaction
The pointer() callback gets its value by reading the I2S BYTE_COUNT
register.  This is a 64-bit runnning transaction counter. If a
transaction was aborted in the middle of a sample buffer, the counter will
stop counting on a number divisible by the buffer size.  Since we actually
use it as a pointer into an aligned buffer, however, we do want to ensure
that it always starts at a number divisible by the buffer size when
starting a transaction, hence we reset it whenever starting a transaction.

To accomplish this, it wasn't necessary to zero bytescount at the
termination of each transaction, so remove this unnecessary code.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 1a337a1e7885085d224583c766614e5945bde671
https://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git)

BUG=b:111073426
TEST=(with all patches in series) capture & playback to headphones
 # cat /proc/interrupts | grep ACP_IRQ
 # arecord -D hw:0,1 -f dat -d 1 /tmp/t1.wav
 # cat /proc/interrupts | grep ACP_IRQ
 => Difference between ACP_IRQ interrupt count is ~= 46
 # aplay -D hw:0,0 -f dat /tmp/t1.wav
 => Sound plays out of Headphones
TEST=(with all patches in series) capture from DMIC & playback to speakers
 # cat /proc/interrupts | grep ACP_IRQ
 # arecord -D hw:0,3 -f dat -d 1 /tmp/t1.wav
 # cat /proc/interrupts | grep ACP_IRQ
 => Difference between ACP_IRQ interrupt count is ~= 46
 # aplay -D hw:0,2 -f dat /tmp/t1.wav
 => Sound plays out of speakers

Change-Id: I0b1a657417fc8104930ae5ded338fabbe165c245
Reviewed-on: https://chromium-review.googlesource.com/1123302
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>