CHROMIUM: iwl7000: mvm: reconfigure queues during inactivity check
We currently reconfigure the queues after the inactivity check,
but only in one of the two callers. This might leave queues in
a state where the TID owner is wrong, if called when reserving
a queue for a new station.

Clean this up and do the reconfiguration inside the inactivity
check function. This requires changing the locking, but one of
the two places already holds the mvm mutex and the other easily
can.

Change-Id: I7029e51030bd72862ef9f68b0cd09f965e262f75
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: bfabdb7a7d65462ece6cef38cf6d5c47743e55ec