UPSTREAM: mac80211: add stop/start logic for software txqs
sometimes, it is required to stop the transmissions momentarily and
resume it later; stopping the txqs becomes very critical in scenarios where
the packet transmission has to be ceased completely. for example, during
the hardware restart, during off channel operations,
when initiating csa(upon detecting a radar on the dfs channel), etc.

the tx queue stop/start logic in mac80211 works well in stopping the tx
when drivers make use of netdev queues, i.e, when qdiscs in network layer
take care of traffic scheduling. since the devices implementing
wake_tx_queue can run without qdiscs, packets will be handed to mac80211
directly without queueing them in the netdev queues.

also, mac80211 does not invoke any of the
netif_stop_*/netif_wake_* apis if wake_tx_queue is implemented.
since the queues are not stopped in this case, transmissions can continue
and this will impact negatively on the operation of the wireless device.

for example,
during hardware restart, we stop the netdev queues so that packets are
not sent to the driver. since ath10k implements wake_tx_queue,
tx queues will not be stopped and packets might reach the hardware while
it is restarting; this can make hardware unresponsive and the only
possible option for recovery is to reboot the entire system.

there is another problem to this, it is observed that the packets
were sent on the dfs channel for a prolonged duration after radar
detection impacting the channel closing time.

we can still invoke netif stop/wake apis when wake_tx_queue is implemented
but this could lead to packet drops in network layer; adding stop/start
logic for software txqs in mac80211 instead makes more sense; the change
proposed adds the same in mac80211.

signed-off-by: manikanta pubbisetty <mpubbise@codeaurora.org>
signed-off-by: johannes berg <johannes.berg@intel.com>

(cherry picked from commit 21a5d4c3a45ca608477a083096cfbce76e449a0c)

BUG=b:117853860
TEST=Done network_WiFi_Reset avl test success

Change-Id: I6a44bf56a804b777635dba9613169cc4c2d9ee51
Reviewed-on: https://chromium-review.googlesource.com/1397172
Commit-Ready: Grant Grundler <grundler@chromium.org>
Tested-by: Wen Gong <wgong@qti-qualcomm.corp-partner.google.com>
Reviewed-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Wen Gong <wgong@qti-qualcomm.corp-partner.google.com>
Reviewed-by: Brian Norris <briannorris@chromium.org>