UPSTREAM: amd/dc/dce100: On dce100, set clocks to 0 on suspend
[Why]
When a dce100 asic was suspended, the clocks were not set to 0.
Upon resume, the new clock was compared to the existing clock,
they were found to be the same, and so the clock was not set.
This resulted in a pernicious blackscreen.

[How]
In atomic commit, check to see if there are any active pipes.
If no, set clocks to 0

BUG=b:112542653
TEST=boot Grunt, run graphics_SanAngeles and login_LoginSuccess

Signed-off-by: David Francis <David.Francis@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit 8d4235f71513cdccd9dc52b674323c3591552bc1)
Signed-off-by: Dominik Behr <dbehr@chromium.org>

Change-Id: Ic0f4dcdf635e454d6740657668c8aef97e8b119c