FROMLIST: ASoC: AMD: Make ACP->SYSMEM DMA non circular
In capture case we don't want ACP to SYSMEM dma
to be circular. This is because if an in place DSP
filter is applied to captured output then circular DMA
can overwrite the filter value with stale data.

Signed-off-by: Akshu Agrawal <akshu.agrawal@amd.com>

(am from https://patchwork.kernel.org/patch/10556579/)

BUG=b:110286632
TEST= audio_LoopbackLatency passes
no noise in an hour of hangout w/ CL:1124875

Change-Id: I8cb94e08819169e47e9c5a58e032bb2b796572e7
Reviewed-on: https://chromium-review.googlesource.com/1141673
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>