UPSTREAM: drm/msm/gpu: Capture the state of the GPU
Add the infrastructure to capture the current state of the GPU and
store it in memory so that it can be dumped later.

For now grab the same basic ringbuffer information and registers
that are provided by the debugfs 'gpu' node but obviously this should
be extended to capture a much larger set of GPU information.

Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Rob Clark <robdclark@gmail.com>
(cherry picked from commit e00e473d9817e03cddbaf181a491c42ae8373482)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:80589817
TEST=Tested on cheza

Change-Id: I99625e4acb3ee27ed01b4f7d11f768caee37f275
Reviewed-on: https://chromium-review.googlesource.com/1158919
Commit-Ready: Sean Paul <seanpaul@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Kristian H. Kristensen <hoegsberg@chromium.org>