FROMLIST: ASoC: SOF: modify load_firmware() to indicate first boot
Modify the signature for load_firmware to add a flag indicating
if this is the first time the DSP is being booted up.
Based on the platform specific implementation, this flag can be
used to decide whether the firmware can be booted from memory
in self-retention or to be downloaded again.

Note that this patch does not change the implementation of any
platform specific load_firmware() method. It merely adds a flag
that can be used to modify the implementation later if needed.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
(cherry picked from commit 4b76c2d8053757857666d002df1e7525fd2f14ee)
Link: https://github.com/thesofproject/linux/commit/4b76c2d8053757857666d002df1e7525fd2f14ee
Signed-off-by: Samaga Prasanna Krishna <samaga.krishna@intel.corp-partner.google.com>

BUG=b:110025789
TEST=Test Audio use cases after full SOF series patches are applied.

Change-Id: I4f343b461b5d17ed16c45ddea418c79c138f7229
Reviewed-on: https://chromium-review.googlesource.com/1171157
Commit-Ready: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>