CHROMIUM: iwl7000: only use hw_rf_id when mvm is compiled
If MVM is not enabled, we don't read the hw_rf_id, but try to use it
later anyway.  Normally it works because in non-mvm cfg's the rf_id
remains set to 0, but it is wrong.

To solve this, move the block that accesses hw_rf_id into the ifdef.

Change-Id: Ibefb444e6a738e3690b28466bc65a0acbe62de5e
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 99ba476b70b99d6489323088b5b502860e66fc5d