CHROMIUM: iwl7000: fix cfg structs for 22000 with different RF modules
We have to choose different configuration and different firmwares
depending on the external RF module that is installed.  Since the
external module is not represented in the PCI IDs, we need to change
the configuration at runtime, after checking the RF ID of the module
installed.  We have a bit of a mess in the code that does this,
because it applies cfg's according to the RF ID only, ignoring the
integrated module that is in use.

Fix that for some devices by adding correct configurations for them
and not ignoring the integrated module's type when making the
decision.

Change-Id: I170f7dfe2baff54e5184fa605eb1497145d0903d
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: c7ca889bb0b3c1c94a47d66fd87e4e7538f44b22