CHROMIUM: iwl7000: move iwl_nvm_check_version() into dvm
This function is only half-used by mvm (i.e. only the nvm_version part
matters, since the calibration version is irrelevant), so it's
pointless to export it from iwlwifi.  If mvm uses this function, it
has the additional complexity of setting the calib version to a bogus
value on all cfg structs.

To avoid this, move the function to dvm and make a simple comparison
of the nvm_version in mvm instead.

Change-Id: Ic3e04a5f887f12684d7eb775586c5801c9d765e2
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: adbf81aa024ec1608555d4ebf8d4ad8856045b61