CHROMIUM: iwl7000: pcie: use iwl_tx_cmd_gen2 and not iwl_tx_cmd
When building AMSDU for gen2, code uses iwl_tx_cmd. The only
updated field is len, which is in the same location, so it
is not a bug. However, it is a bit confusing and error prone,
so change it.

Change-Id: I4d9748075bd69b1ad7971b9cda2170a77c93608c
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
iwl7000-tree: f51cbdf1928a1bba0f4739f2db0d10aec6dfa7f7