CHROMIUM: iwl7000: pcie: lock txq a bit later in reclaim code
In reclaim code, we don't need to take the queue lock for
waking the queue. The code section is executed only when
the tx path is stopped, and since the reclaim path is not
executed in parallel to itself, no one can update the queue
pointers, and accessing them is safe without a lock.

Change-Id: I33227111eb13f9e12d18c357145f519b887d3a51
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
iwl7000-tree: 659fb08cec140d2d95100196c8801f5eb3d1bf2c