CHROMIUM: arm64: dts: qcom: sdm845-cheza: Update ldo19 and 28
This change corrects the voltage for LDOs 19 and 28. These rails
go up to, and should be at, 3.3V, however AOP firmware was limiting
them to 3.0V. Since setting the regulators to 3.3V with old firmware
causes the system not to boot, adjust the voltage back down for rev0-2.
Otherwise everyone has to go on the same debugging odyssey.

BUG=b:112920779
TEST=Boot Cheza r2 with updated AOP firmware, r1 without

Signed-off-by: Evan Green <evgreen@chromium.org>

Change-Id: Idfb7b1be52e74f1a94475b3614df679c54092495
Reviewed-on: https://chromium-review.googlesource.com/1197704
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>