CHROMIUM: arm64: dts: qcom: sdm845-cheza: No more -rev0
The -rev0 hardware was pretty broken and continuing to support it is
more effort than it's worth.  Kill support.

NOTE: for now we'll leave a stub device tree file there with a warning
message in the model field.  This way if someone tries to put a new
kernel on an old -rev0 they'll get a message that will help them
realize that they're unsupported.  Without this the BIOS will just
pick some other cheza device tree and they'll probably get weird
failures.

BUG=b:114019802
TEST=Hack it so -rev1 picks -rev0 dts and see it not boot.

Change-Id: I49d9311d81d8baa0c100f01bd6571ccb9d66218a
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1211686
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>