CHROMIUM: arm64: dts: qcom: Switch edp regulator to load switch
The plan to use LDO19 from the PMIC is not so great because we can only
really achieve 3.0V instead of 3.3V at the panel. Let's go back to the
old way of doing things and use the regualtor load switch from the 3.3V
supply. This will allow us to get a proper 3.3V here and not sit on the
margin of the display panel spec.

TEST=boot up rev3, see 3.3V at panel pin
BUG=None

Change-Id: Ie1333c7d59dee47816fcbb02155a5f39fc023bec
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1216262
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>