CHROMIUM: iwl7000: chromeOS: also use our hardcoded WMM rules on chromeos-4.4
Kirtika backported the ETSI WMM implementation to chromeos-4.4 and
initially we wanted to use that implementation on chromeos-4.4 instead
of ours.  But it was not fully working because in mac80211.c we were
checking for >= 4.17 anyway, so the code needed there was not being
used in 4.4.

Fix this by always using our hardcoded WMM rules on anything < 4.17,
including 4.4.

Change-Id: Ib4955b8ded9f0da565762a4a60b41e9d820e3966
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: b9ff56766a31d47d38e3c22416683d768c567d5a