UPSTREAM: drm/amdkfd: Change the control stack MTYPE from UC to NC on GFX9
[ Upstream commit 15426dbb65c5b37680d27e84d58a1ed3b8532518 ]

CWSR fails on Raven if the control stack is MTYPE_UC, which is used
for regular GART mappings. As a workaround we map it using MTYPE_NC.

The MEC firmware expects the control stack at one page offset from the
start of the MQD so it is part of the MQD allocation on GFXv9. AMDGPU
added a memory allocation flag just for this purpose.

BUG=b:112542653
TEST=boot Grunt, run graphics_SanAngeles and login_LoginSuccess

Acked-by: Alex Deucher <alexander.deucher@amd.com>
Signed-off-by: Yong Zhao <yong.zhao@amd.com>
Reviewed-by: Felix Kuehling <Felix.Kuehling@amd.com>
Signed-off-by: Felix Kuehling <Felix.Kuehling@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
Signed-off-by: Sasha Levin <alexander.levin@microsoft.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit c2fd5b21cbd9fc777fde1509a3e709450094b1ba)
Signed-off-by: Dominik Behr <dbehr@chromium.org>

Change-Id: I7238ebd5f7a1e04982f9a59c0ebc613a6867b97c