UPSTREAM: drm/amdgpu: fix the page fault of raven2
While the apg_end address is 0xffffffff, if add 1 with it, the value will be
overflow and roll back to 0. So when 0 is written to
mmMC_VM_SYSTEM_APERTURE_HIGH_ADDR, the system aperture is actually disabled. And
so any access to vram will trigger a page fault.

Raven2's HW issue only need increase the vram end address, and needn't do it on
the agp.

BUG=b:136019532
TEST=boot grunt, octopus, run login_LoginSuccess, cheets_StartAndroid

Signed-off-by: Huang Rui <ray.huang@amd.com>
Acked-by: Alex Deucher <alexander.deucher@amd.com>
Reviewed-by: Christian König <christian.koenig@amd.com>
Cc: Marek Olšák <marek.olsak@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit 75986276217b8fcb4d5aa8fb76540fbb26c061de)
Signed-off-by: Dominik Behr <dbehr@chromium.org>

Change-Id: I06ff5fa6aacaf5177863bd4b81d9107fa41aae3b