FROMLIST: ASoC: SOF: use suspend_late/resume_early instead of suspend/resume
The normal behaviour for PCI devices is that when the system suspend
occurs while the device is runtime suspended, its runtime_resume
method gets called to bring it back to full power. This seems to
cause race condition between runtime_resume and system suspend
and prevents the system from entering S2.

Also, it is recommended that bus controllers be suspended in the late
suspend phase and resumed in the resume early phase. This ensures that
the runtime pm for the devices is disabled before the suspend_late
is called. Therefore, switch to using suspend_late and resume_early
for the SOF PCI device.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Signed-off-by: Harshapriya N <harshapriya.n@intel.com>
(cherry picked from commit 6cbb72b99333484b99f3e7f854a2af44830fb49d)
Link:https://github.com/thesofproject/linux/commit/6cbb72b99333484b99f3e7f854a2af44830fb49d

BUG=b:110025789
TEST=Test Audio use cases after full SOF series patches are applied.

Change-Id: If574527f788c64003d0c8e78ef7f9c0d5c7312d8
Signed-off-by: Naveen Manohar <naveen.m@intel.com>
Signed-off-by: Samaga Prasanna Krishna <samaga.krishna@intel.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/1242636
Commit-Ready: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>