CHROMIUM: Bluetooth: make splitter switchable
Adds a sysfs file that can be used to enable or disable HCI splitter. By
default the splitter is in an uninitialized state where it will allow
nobody access to the bluetooth chip. It can be once per boot moved to an
enabled or a disabled state by writing to a sysfs file which can be
found at /sys/devices/virtual/misc/hci_le/le_splitter_enabled. The
acceptable values to write are standard kernel sysfs boolean values, eg
"Y" or "yes" or "1", etc... Once a transition to one of the states
happens, any further attempts to make another one will be denied. The
current state can be read out of the same file and will always report
"NOT SET\n", "ON\n", or "OFF\n". Since bluez will be mightily confused
about not being able to talk to the chip at boot, it will give up on
it. So after moving to the enabled or disabled state, it is advised to
kill and restart the userspace bluetoothd daemon, and then tell the
kernel to bring up the interface again via the command
"hciconfug hci0 up" or its equivalent. With the splitter on, newblue
can be run, and with the splitter off, bluez has access to LE and
newblue will fail to initialize if such initilization is attempted.

BUG=b:74813838
TEST=manual for all 3 states. all works as expected

Change-Id: I47d09cc423407cdb239a2428bbcff59773ecf021
Signed-off-by: Dmitry Grinberg <dmitrygr@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1551949
Commit-Ready: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>