UPSTREAM: mac80211: avoid reflecting frames back to the client
I'm not really sure exactly _why_ I've been carrying a note
for what's probably _years_ to check that we don't do this,
but we clearly do reflect frames back to the station itself
if it sends such.

One way or the other, it's useless since the station doesn't
really need the AP to talk to itself, so suppress it.

While at it, clarify some of the logic by removing skb->data
references in favour of the destination address (pointer) we
already have separately.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 42dca5ef24fc846ff133796980e21e5788ddfb70)

BUG=none
TEST=`tast run -var=network.HostapHwsim.runArgs='ap_open_no_reflection'
     localhost:9222 network.HostapHwsim.full`

Signed-off-by: Brian Norris <briannorris@chromium.org>
Change-Id: I32d92dcbe6baec19d4700b701964924486dd6541
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2185392
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>