CHROMIUM: security: Add nanopb support
Add the nanopb engine and use a custom header (pbsystem.h) to make it
work in the Linux kernel. The nanopb license is included in the
dedicated directory.

Github depot: https://github.com/nanopb/nanopb

The code implementation was not changed, instead build flags are used
to build it without libc dependencies. This code doesn't respect Linux
as it generates code and changing the implementation will be a significant
amout of work for any proto change. Instead, a version will be proposed
upstream for audit record binary serialization and other purposes.

BUG=chromium:918980
TEST=Build, boot and GCP internal testing.

Signed-off-by: Thomas Garnier <thgarnie@google.com>
Change-Id: Ida462a6d4731fd038aa5b8ebde0b1561c56ee060
Reviewed-on: https://chromium-review.googlesource.com/1379658
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Robert Kolchmeyer <rkolchmeyer@google.com>